requirements.txt
transformers
torch
torchvision 
torchaudio
pinecone-client